system(U5) update STM32U5xx HAL Drivers to v1.1.0
Included in STM32CubeU5 FW v1.1.0

+++ system/Drivers/STM32U5xx_HAL_Driver/Inc/stm32u5xx_hal_uart_ex.h
@@ -296,72 +296,72 @@ HAL_StatusTypeDef HAL_UARTEx_ClearConfigAutonomousMode(UART_HandleTypeDef *huart
   * @retval UART clocking source, written in __CLOCKSOURCE__.
   */
 #if defined(USART6)
-#define UART_GETCLOCKSOURCE(__HANDLE__,__CLOCKSOURCE__) \
-  do {                                                  \
-    if((__HANDLE__)->Instance == USART1)                \
-    {                                                   \
-      (__CLOCKSOURCE__) = RCC_PERIPHCLK_USART1;         \
-    }                                                   \
-    else if((__HANDLE__)->Instance == USART2)           \
-    {                                                   \
-      (__CLOCKSOURCE__) = RCC_PERIPHCLK_USART2;         \
-    }                                                   \
-    else if((__HANDLE__)->Instance == USART3)           \
-    {                                                   \
-      (__CLOCKSOURCE__) = RCC_PERIPHCLK_USART3;         \
-    }                                                   \
-    else if((__HANDLE__)->Instance == UART4)            \
-    {                                                   \
-      (__CLOCKSOURCE__) = RCC_PERIPHCLK_UART4;          \
-    }                                                   \
-    else if((__HANDLE__)->Instance == UART5)            \
-    {                                                   \
-      (__CLOCKSOURCE__) = RCC_PERIPHCLK_UART5;          \
-    }                                                   \
-    else if((__HANDLE__)->Instance == USART6)           \
-    {                                                   \
-      (__CLOCKSOURCE__) = RCC_PERIPHCLK_USART6;         \
-    }                                                   \
-    else if((__HANDLE__)->Instance == LPUART1)          \
-    {                                                   \
-      (__CLOCKSOURCE__) = RCC_PERIPHCLK_LPUART1;        \
-    }                                                   \
-    else                                                \
-    {                                                   \
-      (__CLOCKSOURCE__) = 0U;                           \
-    }                                                   \
+#define UART_GETCLOCKSOURCE(__HANDLE__,__CLOCKSOURCE__)           \
+  do {                                                            \
+    if((__HANDLE__)->Instance == USART1)                          \
+    {                                                             \
+      (__CLOCKSOURCE__) = (uint32_t)RCC_PERIPHCLK_USART1;         \
+    }                                                             \
+    else if((__HANDLE__)->Instance == USART2)                     \
+    {                                                             \
+      (__CLOCKSOURCE__) = (uint32_t)RCC_PERIPHCLK_USART2;         \
+    }                                                             \
+    else if((__HANDLE__)->Instance == USART3)                     \
+    {                                                             \
+      (__CLOCKSOURCE__) = (uint32_t)RCC_PERIPHCLK_USART3;         \
+    }                                                             \
+    else if((__HANDLE__)->Instance == UART4)                      \
+    {                                                             \
+      (__CLOCKSOURCE__) = (uint32_t)RCC_PERIPHCLK_UART4;          \
+    }                                                             \
+    else if((__HANDLE__)->Instance == UART5)                      \
+    {                                                             \
+      (__CLOCKSOURCE__) = (uint32_t)RCC_PERIPHCLK_UART5;          \
+    }                                                             \
+    else if((__HANDLE__)->Instance == USART6)                     \
+    {                                                             \
+      (__CLOCKSOURCE__) = (uint32_t)RCC_PERIPHCLK_USART6;         \
+    }                                                             \
+    else if((__HANDLE__)->Instance == LPUART1)                    \
+    {                                                             \
+      (__CLOCKSOURCE__) = (uint32_t)RCC_PERIPHCLK_LPUART1;        \
+    }                                                             \
+    else                                                          \
+    {                                                             \
+      (__CLOCKSOURCE__) = 0U;                                     \
+    }                                                             \
   } while(0U)
 #else
-#define UART_GETCLOCKSOURCE(__HANDLE__,__CLOCKSOURCE__) \
-  do {                                                  \
-    if((__HANDLE__)->Instance == USART1)                \
-    {                                                   \
-      (__CLOCKSOURCE__) = RCC_PERIPHCLK_USART1;         \
-    }                                                   \
-    else if((__HANDLE__)->Instance == USART2)           \
-    {                                                   \
-      (__CLOCKSOURCE__) = RCC_PERIPHCLK_USART2;         \
-    }                                                   \
-    else if((__HANDLE__)->Instance == USART3)           \
-    {                                                   \
-      (__CLOCKSOURCE__) = RCC_PERIPHCLK_USART3;         \
-    }                                                   \
-    else if((__HANDLE__)->Instance == UART4)            \
-    {                                                   \
-      (__CLOCKSOURCE__) = RCC_PERIPHCLK_UART4;          \
-    }                                                   \
-    else if((__HANDLE__)->Instance == UART5)            \
-    {                                                   \
-      (__CLOCKSOURCE__) = RCC_PERIPHCLK_UART5;          \
-    }                                                   \
-    else if((__HANDLE__)->Instance == LPUART1)          \
-    {                                                   \
-      (__CLOCKSOURCE__) = RCC_PERIPHCLK_LPUART1;        \
-    }                                                   \
-    else                                                \
-    {                                                   \
-      (__CLOCKSOURCE__) = 0U;                           \
-    }                                                   \
+#define UART_GETCLOCKSOURCE(__HANDLE__,__CLOCKSOURCE__)           \
+  do {                                                            \
+    if((__HANDLE__)->Instance == USART1)                          \
+    {                                                             \
+      (__CLOCKSOURCE__) = (uint32_t)RCC_PERIPHCLK_USART1;         \
+    }                                                             \
+    else if((__HANDLE__)->Instance == USART2)                     \
+    {                                                             \
+      (__CLOCKSOURCE__) = (uint32_t)RCC_PERIPHCLK_USART2;         \
+    }                                                             \
+    else if((__HANDLE__)->Instance == USART3)                     \
+    {                                                             \
+      (__CLOCKSOURCE__) = (uint32_t)RCC_PERIPHCLK_USART3;         \
+    }                                                             \
+    else if((__HANDLE__)->Instance == UART4)                      \
+    {                                                             \
+      (__CLOCKSOURCE__) = (uint32_t)RCC_PERIPHCLK_UART4;          \
+    }                                                             \
+    else if((__HANDLE__)->Instance == UART5)                      \
+    {                                                             \
+      (__CLOCKSOURCE__) = (uint32_t)RCC_PERIPHCLK_UART5;          \
+    }                                                             \
+    else if((__HANDLE__)->Instance == LPUART1)                    \
+    {                                                             \
+      (__CLOCKSOURCE__) = (uint32_t)RCC_PERIPHCLK_LPUART1;        \
+    }                                                             \
+    else                                                          \
+    {                                                             \
+      (__CLOCKSOURCE__) = 0U;                                     \
+    }                                                             \
   } while(0U)
 #endif /* USART6 */
